Dhivya's starter tasks: Added Datacard and Modelcard
My experience creating datacards and modelcards was very unique and interesting. I created a  datacard for the GTSRB dataset and created a model card for a CNN model which I deployed on my own. This experience helped me understand the documentation process and helped me realize how much transparency is important. The model lacked various aspects of fairness and, most notably, associated risks, impact, and explanations for the prediction. The challenges I faced were finding out about the access, retention and the wipeout of the dataset. I used the dataset documentation website and the research paper published for documenting the data and model cards. I look forward to contributing many more data and model cards.