Add Interactive Calculators to new Labs directory
Woo woo, we are moving to the main branch! The interactive calculator folder has been moved from the personal/child repo to the parent repo. 

We built this with SvelteKit and PostCSS. Please refer to the documents, such as the blog, design process (Figma), and API/Build/Contribute in our README.md and /docs for more information. 

Note: The following has known issues (refer to /docs), such as migrating/breaking changes in SvelteKit. 

Since it's an external application/embedded into the website, it does not have any known crashes or potential to break the build of Playbooks Lab. 

Currently for preproduction uses. Uploaded publicly for future pull requests and more.